Basic cleanup
Improving word choice here and there. Also I'm wondering if I can rewrite more of this page to explain a little better?

Example for the intro: "With Fork CMS it's easy to customize your site's design and layout with either a pre-built or a custom theme. On the [Extensions](http://www.fork-cms.com/extensions/themes) page you'll find a list of themes available for download. These can easily be downloaded and installed on your own Fork CMS installation."